--- src/components/mouse-stalker/index.ts
@@ -1,1 +1,0 @@
-export { MouseStalker } from './mouse-stalker'
